+++ gulpfile.js
@@ -178,6 +178,7 @@ gulp.task('make-package', function() {
     }
 
     // if we are on OSX make sure to use gtar for compatibility with Linux
+    // otherwise we see lots of error message when extracting with GNU tar
     var tar = /^darwin/.test(process.platform) ? 'gtar' : 'tar';
 
     var tarResult = spawnSync(tar, [